v0.2.8 (#10)
* Support `unfreeze` operation.

* New `mutable` and `immutable` context managers that turn on and turn off mutability.

* Pax now makes `deepcopy` of `not_tree` attributes when `immutable` mode is on.

* Add `pax.jit`, `pax.grad`, `pax.vmap`, `pax.pmap`, which are thin wrapper of `jax.*`. These wrappers help to turn on pax's strict mode.

* Fix DCGAN data link.

* Pax now detects MODULE_SUBTREE attributes automatically. No need to use `self.register_module_subtree` any more.

* Don't allow to assign empty list, or empty dict to an attribute of a Module to prevent bugs.

* Add `assertStructureEqual` method that print out the difference between two modules.

* New immutability and limitations documentations.

+++ tests/test_counter.py
@@ -10,7 +10,7 @@ def test_counter():
 
         def __init__(self, start_value: int = 0):
             super().__init__()
-            self.register_state("counter", jnp.array(start_value))
+            self.register_state("counter", jnp.array(start_value, dtype=jnp.int32))
             self.register_parameter("bias", jnp.array(0.0))
 
         def __call__(self, x):
